Updated ResultOutput methods to use get_file instead of get<string> to avoid edgecase where the response is parsed as something other than a string.
