+++ package.json
@@ -15,16 +15,16 @@
   },
   "devDependencies": {
     "@types/jest": "^23.3.1",
+    "@types/node": "^10.17.54",
     "jest": "^23.4.2",
     "ts-jest": "^23.0.1",
     "tslint": "^5.11.0",
     "tslint-consistent-codestyle": "^1.13.3",
     "tslint-eslint-rules": "^5.3.1",
-    "typescript": "^3.9.7"
+    "typescript": "~3.9.3"
   },
   "dependencies": {
-    "@types/node": "^10.17.49",
-    "axios": "^0.21.1"
+    "axios": "^0.18.1"
   },
   "main": "dist/index.js",
   "types": "dist/index.d.ts"
